Improve code coverage (#5)
Improve code coverage

Few things were implemented to achieve that:
- Remove err checks where a nil err is always returned
- Improved error reporting
- Data validation

Error handling was removed for Builder.Write() and Builder.WriteString(). The documentation states that both functions return nil. It was also removed for json.Marshal(Leg). The documentation states that json.Marshal() only returns an error if v is a channel, complex, or function value. Leg is a data model and would only contain built-in types.

Sentinel errors were replaced with DecodeError to provide a richer and more informative error report. The error report includes the boarding pass data, the position in the string where the error occurred, the item that was processed, and the reason why the error occurred. Data validation was also implemented as an extension of error reporting.